Fix the build with _FORTIFY_SOURCE enabled
util.h uses various HAVE_ macros that are defined in config.h.
Currently everyone needs to include it beforehand, but it's much safer
to just include it directory in utils.h to avoid some declarations that
may conflict with system headers.

This fixes both _FORTIFY_SOURCE and an inconsistent declaration of
pipe2(2) that I hit while trying to build libudev-devd originally.